Bug Fix: Dereference & create mutable reference
-  Root cause was the set() of impl of Utf8PtrData.
Apparently, mut p was being assigned to (*self.param)
so when the compiler tries to update the return_size
based on the new provided value, it reflects in mut p
but it is not being propagated to ossl_param.

- Tried let p = unsafe{&mut *self.param} instead and test passed.
It will provide the mutable reference to the original ossl_param
instead of the copy p.

- References/further read:
https://users.rust-lang.org/t/when-does-dereferencing-a-pointer-create-a-copy/51955
https://doc.rust-lang.org/reference/expressions.html#place-expressions-and-value-expressions
https://doc.rust-lang.org/reference/expressions/block-expr.html
https://bluss.github.io/rust/fun/2015/10/11/stuff-the-identity-function-does/