hack(bindings.rs): disable function type check
The check_dispatch_table_entry_type function was commented out and
replaced with a phantom statement using an Option.

This was necessary as sometime we need to evaluate this in a `const`
context and the current version of the ceck does not evaluate correctly.

See https://mattermost.nisec.fi/nisec/pl/gn8p861qsfbyud8ifngmseaawo