Implement an Iterator that gives OSSLParam structs from a OSSL_PARAM pointer
Now if you have `params: *const OSSL_PARAM` or `params: *mut OSSL_PARAM`,
you can do `let params_iter = OSSLParamIter::new(params)` and then do
things like `for p in params_iter` or call methods like `filter()` and
`map()` on the iterator. Unfortunately since the raw pointer is a
foreign type we can't implement an `iter()` method on it afaik, so
hopefully this type of design is usable as-is.